// src/lib/api/generated/models/remoteHttpMethod.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Hubuum REST API
 * OpenAPI documentation for the Hubuum REST service.
 * OpenAPI spec version: 0.0.1
 */

export type RemoteHttpMethod = typeof RemoteHttpMethod[keyof typeof RemoteHttpMethod];


export const RemoteHttpMethod = {
  get: 'get',
  post: 'post',
  patch: 'patch',
  delete: 'delete',
} as const;